Fixed code errors for new memory functions and using type long instead of int

+++ include/core/inputter_outputter.h
@@ -6,17 +6,17 @@
 class InputterOutputter
 {
 private:
-    std::vector<int> m_values;
+    std::vector<long> m_values;
     int m_get_position;
     int size();
 public:
     InputterOutputter();
     ~InputterOutputter();
-    void add(int value);
-    void add(std::vector<int> values);
-    int getNext(int * result);
-    int getPosition(int position, int * result);
-    int replace(int value, int position);
+    void add(long value);
+    void add(std::vector<long> values);
+    int getNext(long * result);
+    int getPosition(int position, long * result);
+    int replace(long value, int position);
     bool hasAvailableElements();
 };
