+++ src/test/test_operation_addition.cpp
@@ -19,7 +19,7 @@ void testAddition()
     std::cout << "Running " << addition->getName() << " with opCode " << addition->getOpcode() << std::endl;
     int res = addition->performOperation(mem, 0, addition->getOpcode(), &new_ip, NULL, NULL);
     std::cout << "performOperation Result is " << res << std::endl;
-    int actual_res=0;
+    long actual_res=0;
     mem->getImmediateMode(res_position, &actual_res);
     if (expected_res == actual_res)
     {
